Imaging Sub and Example Windows Clarification
This is a subrule in Windows Clarification to detect and provide safe passage temporarily for devices identified as Imaging.  A label is applied with a 'quasi' time limit, with the applied label being deleted by the delete Label action, i found this to be a cleaner way for audit trails and host logs tracking stand point than an expiring assigned label.